+++ R/build-home.R
@@ -89,8 +89,14 @@
 #'
 #' @section Badges:
 #' Status badges are displayed in the sidebar under the section "Dev status".
-#' This section is automatically populated if the first paragraph of the
-#' homepage consists solely of status badges as linked images.
+#' This section is automatically populated if there is an identifiable badges
+#'  paragraph in the homepage source (index.Rmd, index.md, README.Rmd,
+#'  README.md):
+#'  * pkgdown first looks for a paragraph starting with `<!-- badges: start -->`
+#' and ending with `<!-- badges: end -->` as created by
+#' `usethis::use_readme_md()` or `usethis::use_readme_rmd()`;
+#'  * failing that, pkgdown looks at the first paragraph, and uses it if it
+#'   only contains images.
 #'
 #' @inheritParams build_articles
 #' @export