+++ tests/testthat/test-html-tweak.R
@@ -162,9 +162,27 @@ test_that("badges can't contain an extra text", {
   )
 })
 
+
 test_that("badges can be in special div", {
   expect_equal(
     badges_extract_text('<p></p><div id="badges"><a href="x"><img src="y"></a></div>'),
     '<a href="x"><img src="y"></a>'
   )
+})
+
+test_that("badges-paragraph a la usethis can be found", {
+  string <- '
+  <blockquote>
+  <p>Connect to thisisatest, from R</p>
+  </blockquote>
+  <!-- badges: start -->
+  <p><a href="https://travis-ci.org/thisisatest/thisisatest"><img src="https://travis-ci.org/thisisatest/thisisatest.svg?branch=master" alt="Linux Build Status"></a> <!-- badges: end --></p>
+  <div id="introduction" class="section level2">
+  <h2 class="hasAnchor">
+  <a href="#introduction" class="anchor"></a>Introduction</h2>
+  <p>The thingie is a blabla.</p>
+  <p>The <code>thisisatest</code> package also blabla.</p>'
+
+  badges_page <- xml2::read_html(string)
+  expect_equal(length(badges_extract(badges_page)), 1)
 })